+++ docs/Whats-New.md
@@ -221,6 +221,15 @@ SaveGameOnScenarioStart=true     ; boolean
   10103=Load Onto Transports,0,0,1,[LONG DESC]
   10104=Chronoshift to Enemy Base,20,0,1,[LONG DESC]
   14004=Force Global OnlyTargetHouseEnemy value in Teams,20,0,1,[LONG DESC]
+  14006=Set House Hate Value Modifier,20,0,1,[LONG DESC]
+  14007=Modify House Hate Using House Index,20,0,1,[LONG DESC]
+  14008=Modify Hate Values From A List Of Countries,28,0,1,[LONG DESC]
+  14009=Modify Hate Value Against A Random Country From A List Of Countries,28,0,1,[LONG DESC]
+  14010=Set The Most Hated House ("<" Comparison),20,0,1,[LONG DESC]
+  14011=Set The Most Hated House (">" Comparison),20,0,1,[LONG DESC]
+  14012=Set The Most Hated House Randomly,0,0,1,[LONG DESC]
+  14013=Reset Hate Against Other Houses,0,0,1,[LONG DESC]
+  14014=Set A House As The Most Hated House Of The Map,20,0,1,[LONG DESC]
   16006=Set House Index For Managing AI Triggers,20,0,1,[LONG DESC]
   16007=Enable Or Disable All AI Triggers,21,0,1,[LONG DESC]
   16008=Enable AI Triggers From List,28,0,1,[LONG DESC]
@@ -310,10 +319,12 @@ SaveGameOnScenarioStart=true     ; boolean
   27=Global variables,-6
   28=AI Scripts List, -7
   29=AI Target Type,-8
+  28=AI Houses List, -7
 
   [ScriptParamTypes]
   7=AIScriptsList,1,1,0
   8=AITargetTypes,1,1,0
+  7=AIHousesList,1,1,0
   ```
 </details>
 
@@ -674,6 +685,7 @@ New:
 - Script action to regroup temporarily around the Team Leader (by FS-21)
 - Script action to randomly skip next action (by FS-21)
 - Script action for timed script action jumps (by FS-21)
+- Script action for modifying AI anger against other houses (by FS-21)
 - ObjectInfo now shows current Target and AI Trigger data (by FS-21)
 - Shield absorption and passthrough customization (by Morton)
 - Limbo Delivery of buildings (by Morton)